added main content
This repository contains a blank postgreSQL schema for storing
concurrently collected behavioral and neural data. Or just behavioral
data. Or just neural data.

It was not initially intended for sharing, and as such it is a bit messy
and hacked-together. That being said, I think it will provide a valuable
starting place for a cleaner, more comprehensive database schema.